# contracts/kawsay-estancias-openapi.yaml
...
openapi: 3.0.3
info:
  title: Kawsay Estancias API — contrato propio
  version: 1.2.0
  description: "Contrato propio del dominio de alojamientos de Kawsay Estancias.
    Es la fuente de esquemas para validación AJV y documentación del núcleo
    REST. Conserva nombres de operaciones compatibles con la plantilla
    académica, con modelos y reglas del proyecto: Ecuador, USD, reservas sin
    solapamientos, código de reserva, JWT revocable, idempotencia y outbox. Las
    extensiones administrativas se documentan adicionalmente mediante DTO de
    NestJS. No hay cobros reales ni entrega automática de webhooks."
  x-origin: Diseñado para Kawsay Estancias; la plantilla se conserva por separado
    como referencia académica.
servers:
  - url: /api/v1
    description: API del despliegue desde el que se abre el contrato
  - url: https://booking-prototipo-alojamientos.vercel.app/api/v1
    description: Producción actual de Kawsay Estancias
security:
  - bearer: []
  - cookie: []
...
paths:
...
  /orders/{orderId}:
    get:
      operationId: get_orders__orderId_
      tags:
        - Reservas
      summary: Consultar una reserva propia o con rol administrador
      description: ""
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/{orderId}/modify:
    post:
      operationId: post_orders__orderId__modify
      tags:
        - Reservas
      summary: Modificar fechas sin solaparse con otra reserva
      description: ""
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: UUID v4. Repetir la misma clave y cuerpo devuelve la misma reserva;
            cambiar el cuerpo produce 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderModifyRequest"
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/{orderId}/cancel:
    post:
      operationId: post_orders__orderId__cancel
      tags:
        - Reservas
      summary: Cancelar y liberar las noches reservadas
      description: ""
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: UUID v4. Repetir la misma clave y cuerpo devuelve la misma reserva;
            cambiar el cuerpo produce 409.
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
...
  /auth/login:
    post:
      operationId: post_auth_login
      tags:
        - Identidad
      summary: Iniciar sesión con JWT y cookie HttpOnly
      description: ""
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookie:
      type: apiKey
      in: cookie
      name: booking_session
  schemas:
...
    AccommodationsGuests:
      type: object
      required:
        - number_of_adults
        - number_of_rooms
      properties:
        number_of_adults:
          type: integer
          minimum: 1
          maximum: 100
        number_of_rooms:
          type: integer
          minimum: 1
        children:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 17
...
    CustomerDetails:
      type: object
      required:
        - first_name
        - last_name
        - email
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
...
    OrderModifyRequest:
      type: object
      properties:
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          $ref: "#/components/schemas/AccommodationsGuests"
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 128
          writeOnly: true
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - customer
    AuthResponse:
      type: object
      required:
        - user
        - access_token
        - token_type
        - expires_in
      properties:
        user:
          $ref: "#/components/schemas/User"
        access_token:
          type: string
        token_type:
          type: string
          enum:
            - Bearer
        expires_in:
          type: integer
          example: 3600
...
    Reservation:
      type: object
      required:
        - order_id
        - locator
        - status
        - accommodation_details
        - checkin
        - checkout
        - guests
        - total_price
        - currency
        - creation_date
      properties:
        order_id:
          type: string
          format: uuid
        locator:
          type: string
          pattern: ^BP-[A-F0-9]{8}$
          description: Código de reserva visible en la confirmación y Mis reservas.
          example: BP-41A92454
        status:
          type: string
          enum:
            - CONFIRMED
            - CANCELLED
        accommodation_details:
          type: object
          required:
            - id
          properties:
            id:
              type: integer
              minimum: 1
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          $ref: "#/components/schemas/AccommodationsGuests"
        customer_details:
          $ref: "#/components/schemas/CustomerDetails"
        total_price:
          type: number
          minimum: 0
        currency:
          type: string
          enum:
            - USD
          default: USD
        creation_date:
          type: string
          format: date-time
        estancia_estado:
          type: string
          enum:
            - CONFIRMADA
            - EN_CURSO
            - COMPLETADA
            - CANCELADA
        puede_resenar:
          type: boolean
        _links:
          type: object
          properties:
            self:
              type: string
            modify:
              type: string
            cancel:
              type: string
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            properties: {}
...
  responses:
    ProblemDetails400:
      description: Solicitud inválida según el contrato o las reglas del dominio.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemDetails401:
      description: Sesión ausente o vencida.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemDetails403:
      description: Permisos insuficientes.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ProblemDetails409:
      description: Fechas ocupadas, cotización vencida o clave idempotente reutilizada
        con otro contenido.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...